issue-tracker-idea-plugin-18 Remove CheckTask
- Removed `CheckTask`
- Documented everything
