+++ Plugin/src/com/madrapps/issuetracker/actions/RefreshIssueListAction.java
@@ -13,6 +13,7 @@ import com.madrapps.issuetracker.listissues.ListIssuesPresenter;
  */
 public class RefreshIssueListAction extends AnAction {
 
+    /** Action ID. This should be the same as in the Plugin xml */
     public static final String ACTION_ID = "IssueTracker.RefreshIssueList";
 
     @Override
