+++ Plugin/src/com/madrapps/issuetracker/actions/ShowDetailsPanelAction.java
@@ -2,7 +2,6 @@ package com.madrapps.issuetracker.actions;
 
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.actionSystem.ToggleAction;
-import com.madrapps.issuetracker.listissues.IListIssuesContract;
 import com.madrapps.issuetracker.listissues.ListIssuesPresenter;
 
 /**
@@ -12,17 +11,16 @@ import com.madrapps.issuetracker.listissues.ListIssuesPresenter;
  */
 public class ShowDetailsPanelAction extends ToggleAction {
 
+    /** Action ID. This should be the same as in the Plugin xml */
     public static final String ACTION_ID = "IssueTracker.ShowDetailsPanel";
 
     @Override
     public boolean isSelected(AnActionEvent e) {
-        final IListIssuesContract.IPresenter presenter = ListIssuesPresenter.getInstance();
-        return presenter.isDetailsPanelShown();
+        return ListIssuesPresenter.getInstance().isDetailsPanelShown();
     }
 
     @Override
     public void setSelected(AnActionEvent e, boolean state) {
-        final IListIssuesContract.IPresenter presenter = ListIssuesPresenter.getInstance();
-        presenter.showDetailsPanel(state);
+        ListIssuesPresenter.getInstance().showDetailsPanel(state);
     }
 }
