+++ Plugin/src/com/madrapps/issuetracker/listissues/IListIssuesContract.java
@@ -30,24 +30,61 @@ public interface IListIssuesContract {
          */
         void updateIssueList(@NotNull List<Task> issuesList, boolean forceUpdate);
 
+        /**
+         * Initializes the ToolWindow components and their actions
+         *
+         * @param project the current project
+         */
         void init(@NotNull Project project);
 
         /**
-         * Show the summary for the selected issue in the summary panel
+         * Show the details for the selected issue in the details panel
+         *
+         * @param description The description of the issue
+         * @param issueUrl    the IssueUrl of the issue
+         * @param comments    the comments of the issue
          */
-        void showSummary(@Nullable String description, @Nullable String issueUrl, @Nullable Comment[] comments);
+        void showDetails(@Nullable String description, @Nullable String issueUrl, @Nullable Comment[] comments);
 
+        /**
+         * Decides whether we should show the Loading component or the Issue Table component
+         *
+         * @param shouldShow true to show the Loading component, false to show the Table component
+         */
         void showLoadingScreen(boolean shouldShow);
 
+        /**
+         * Show the Empty issues list instead of the Table
+         */
         void showEmptyIssueListScreen();
 
+        /**
+         * Opens the issue url in the browser
+         *
+         * @param issueUrl the url of the issue
+         */
         void openInBrowser(@NotNull String issueUrl);
 
+        /**
+         * Show/Hide the details panel
+         *
+         * @param shouldShow if true, show the details panel, if false hide it
+         */
         void showDetailsPanel(boolean shouldShow);
 
+        /**
+         * Get the selected issue in the Issue list table
+         *
+         * @return the selected issue in the issue list, or null if nothing is selected
+         */
         @Nullable
         Task getSelectedIssue();
 
+        /**
+         * Determines if the details panel is shown or not
+         *
+         * @return true if it is shown, false otherwise
+         */
         boolean isDetailsPanelShown();
     }
 
@@ -56,24 +93,77 @@ public interface IListIssuesContract {
      */
     interface IPresenter {
 
+        /**
+         * Pull issues asynchronously from the server
+         *
+         * @param project the current project
+         * @param query   the search query
+         * @param force   if true, force sync all issues, if false adds new issues to existing list
+         */
         void pullIssues(@NotNull Project project, @Nullable String query, boolean force);
 
+        /**
+         * Pull issues asynchronously from the server
+         *
+         * @param project the current project
+         * @param query   the search query
+         * @param offset  the offset (mostly 0)
+         * @param limit   the limit
+         * @param force   if true, force sync all issues, if false adds new issues to existing list
+         */
         void pullIssues(@NotNull Project project, @Nullable String query, int offset, int limit, boolean force);
 
-        void showSummary(@NotNull Task selectedIssue);
+        /**
+         * Show the details of a particular issue
+         *
+         * @param selectedIssue the issue
+         */
+        void showDetails(@NotNull Task selectedIssue);
 
+        /**
+         * Load the issues when the Tool window is first activated
+         *
+         * @param project the current project
+         */
         void loadInitialIssues(@NotNull Project project);
 
+        /**
+         * Sets The ToolWindow view to the presenter, so that the presenter has a reference to it
+         *
+         * @param project the current project
+         * @return the view
+         */
         @NotNull
         IView setView(@NotNull Project project);
 
+        /**
+         * Sets The ToolWindow view to the presenter, so that the presenter has a reference to it
+         *
+         * @param view the view
+         * @return the view
+         */
         @NotNull
         IView setView(@NotNull IView view);
 
+        /**
+         * Opens the issue url in the browser
+         *
+         * @param selectedIssue the issue
+         */
         void openUrl(@NotNull Task selectedIssue);
 
+        /**
+         * Show/Hide the details panel
+         *
+         * @param shouldShow if true show the panel, hide otherwise
+         */
         void showDetailsPanel(boolean shouldShow);
 
+        /**
+         * Determine if the details panel is shown
+         *
+         * @return true if shown, false otherwise
+         */
         boolean isDetailsPanelShown();
     }
 }
