+++ Plugin/src/com/madrapps/issuetracker/listissues/IssuesToolWindowPanel.form
@@ -8,7 +8,7 @@
     <properties/>
     <border type="none"/>
     <children>
-      <grid id="3a4c" binding="mSummaryPanel" layout-manager="GridLayoutManager" row-count="1" column-count="1" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
+      <grid id="3a4c" binding="mDetailsPanel" layout-manager="GridLayoutManager" row-count="1" column-count="1" same-size-horizontally="false" same-size-vertically="false" hgap="-1" vgap="-1">
         <margin top="0" left="0" bottom="0" right="0"/>
         <constraints>
           <grid row="2" column="1" row-span="1" col-span="1" vsize-policy="7" hsize-policy="7" anchor="0" fill="3" indent="0" use-parent-layout="false"/>
@@ -23,7 +23,7 @@
             <properties/>
             <border type="none"/>
             <children>
-              <component id="4c86c" class="javax.swing.JTextPane" binding="mIssueSummaryTextPane">
+              <component id="4c86c" class="javax.swing.JTextPane" binding="mIssueDetailsTextPane">
                 <constraints/>
                 <properties>
                   <background color="-1"/>
