+++ Plugin/src/com/madrapps/issuetracker/listissues/IssuesToolWindowPanel.java
@@ -8,18 +8,14 @@ import com.intellij.openapi.actionSystem.AnAction;
 import com.intellij.openapi.actionSystem.DefaultActionGroup;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.ui.SimpleToolWindowPanel;
-import com.intellij.openapi.util.Comparing;
 import com.intellij.openapi.wm.ToolWindow;
 import com.intellij.tasks.Comment;
 import com.intellij.tasks.Task;
-import com.intellij.tasks.TaskRepository;
 import com.intellij.ui.BrowserHyperlinkListener;
 import com.intellij.ui.PopupHandler;
 import com.intellij.ui.content.Content;
 import com.intellij.ui.content.ContentFactory;
 import com.intellij.ui.table.TableView;
-import com.intellij.util.text.DateFormatUtil;
-import com.intellij.util.ui.ColumnInfo;
 import com.intellij.util.ui.ListTableModel;
 import com.madrapps.issuetracker.actions.OpenIssueInBrowserAction;
 import com.madrapps.issuetracker.actions.RefreshIssueListAction;
@@ -33,104 +29,24 @@ import org.jetbrains.annotations.Nullable;
 
 import java.awt.CardLayout;
 import java.util.ArrayList;
-import java.util.Comparator;
-import java.util.Date;
 import java.util.List;
 import java.util.Locale;
 
-import javax.swing.Icon;
 import javax.swing.JFormattedTextField;
 import javax.swing.JPanel;
-import javax.swing.JTable;
 import javax.swing.JTextPane;
 import javax.swing.ListSelectionModel;
 
+import static com.madrapps.issuetracker.listissues.TableColumns.COLUMN_NAMES;
+import static com.madrapps.issuetracker.listissues.TableColumns.getValueOfDate;
+
 /**
- * This is repsonsible for the ToolWindow GUI. The GUI is backed up be a .form file.
+ * This is responsible for the ToolWindow GUI. The GUI is backed up be a .form file.
  * <p>
  * Created by Henry on 10/19/2016.
  */
 public class IssuesToolWindowPanel extends SimpleToolWindowPanel implements IListIssuesContract.IView {
 
-    private final static ColumnInfo<Task, String> PRESENTABLE_NAME = new ColumnInfo<Task, String>("Name") {
-        public String valueOf(Task object) {
-            return object.getPresentableName();
-        }
-
-        public Comparator<Task> getComparator() {
-            return (o, o1) -> o.getPresentableName().compareTo(o1.getPresentableName());
-        }
-    };
-    private final static ColumnInfo<Task, String> REPOSITORY_TYPE = new ColumnInfo<Task, String>("Repository") {
-        public String valueOf(Task object) {
-            final TaskRepository repository = object.getRepository();
-            if (repository != null) {
-                return repository.getRepositoryType().getName();
-            }
-            return "";
-        }
-
-        public Comparator<Task> getComparator() {
-            return (o, o1) -> {
-                if (o.getRepository() != null && o1.getRepository() != null) {
-                    return o.getRepository().getRepositoryType().getName().compareTo(o1.getRepository().getRepositoryType().getName());
-                } else {
-                    return 0;
-                }
-            };
-        }
-    };
-    private final static ColumnInfo<Task, Icon> ICON = new ColumnInfo<Task, Icon>("") {
-        @Nullable
-        @Override
-        public Icon valueOf(Task task) {
-            return task.getIcon();
-        }
-
-        @Nullable
-        @Override
-        public Comparator<Task> getComparator() {
-            return (o, o1) -> {
-                if (o != null && o1 != null) {
-                    return o.getIcon().toString().compareTo(o1.getIcon().toString());
-                } else {
-                    return 0;
-                }
-            };
-        }
-
-        @Override
-        public Class<?> getColumnClass() {
-            // So that the appropriate Renderer is used to render and Icon
-            return Icon.class;
-        }
-
-        @Override
-        public int getWidth(JTable table) {
-            // So as to prevent users from resizing this column
-            return 50;
-        }
-    };
-    private final static ColumnInfo<Task, String> CREATED_ON = new ColumnInfo<Task, String>("Created On") {
-        public String valueOf(Task object) {
-            return getValueOfDate(object.getCreated());
-        }
-
-        public Comparator<Task> getComparator() {
-            return (o, o1) -> Comparing.compare(o.getCreated(), o1.getCreated());
-        }
-    };
-    private final static ColumnInfo<Task, String> LAST_UPDATED = new ColumnInfo<Task, String>("Last Updated On") {
-        public String valueOf(Task object) {
-            return getValueOfDate(object.getUpdated());
-        }
-
-        public Comparator<Task> getComparator() {
-            return (o, o1) -> Comparing.compare(o.getUpdated(), o1.getUpdated());
-        }
-    };
-    /** The columns of the issues table */
-    private static final ColumnInfo[] COLUMN_NAMES = {ICON, PRESENTABLE_NAME, CREATED_ON, LAST_UPDATED, REPOSITORY_TYPE};
     /** To parse the Markdown text */
     private static final Parser MARKDOWN_PARSER = Parser.builder().build();
     /** To render the markdown as html */
@@ -139,17 +55,26 @@ public class IssuesToolWindowPanel extends SimpleToolWindowPanel implements ILis
     private JPanel mContentPanel;
     /** The table that shows the list of tasks/issues */
     private TableView<Task> mIssuesTable;
-    /** The summary panel that shows the details of an issue when it's selected from the table */
-    private JTextPane mIssueSummaryTextPane;
+    /** The Details panel that shows the details of an issue when it's selected from the table */
+    private JTextPane mIssueDetailsTextPane;
+    /** The toolbar to hold the actions in the ToolWindow */
     private JPanel mToolbar;
+    /** The text field that shows the loading/empty issue message */
     private JFormattedTextField mEmptyMessageTextField;
+    /** Panel that holds the IssuesTable */
     private JPanel mIssuesListPanel;
-    private JPanel mSummaryPanel;
+    /** The Details panel component */
+    private JPanel mDetailsPanel;
     /** The list that's backing up the {@code mIssuesTable} */
     private List<Task> mIssueList;
+    /** The presenter */
     private ListIssuesPresenter mPresenter;
 
-
+    /**
+     * Constructor
+     *
+     * @param toolWindow the ToolWindow from the Factory
+     */
     IssuesToolWindowPanel(ToolWindow toolWindow) {
         super(true, false);
         final Content content = ContentFactory.SERVICE.getInstance().createContent(this, "", false);
@@ -181,11 +106,12 @@ public class IssuesToolWindowPanel extends SimpleToolWindowPanel implements ILis
         initializeComponents();
         initializeActions();
 
+        // Load issues for the first time when the ToolWindow is opened
         mPresenter.loadInitialIssues(project);
     }
 
     @Override
-    public void showSummary(@Nullable String description, @Nullable String issueUrl, @Nullable Comment[] comments) {
+    public void showDetails(@Nullable String description, @Nullable String issueUrl, @Nullable Comment[] comments) {
         final StringBuilder stringBuilder = new StringBuilder();
         stringBuilder.append("<style>")
                 .append(".comment {color: #000; background-color: #ddffff; padding: 5px 10px; border-left: 6px solid #ccc; display:inline}")
@@ -211,7 +137,7 @@ public class IssuesToolWindowPanel extends SimpleToolWindowPanel implements ILis
             stringBuilder.append(String.format(Locale.US, "<a href=%s>%s</a>", issueUrl, issueUrl))
                     .append("<br/>");
         }
-        mIssueSummaryTextPane.setText(stringBuilder.toString());
+        mIssueDetailsTextPane.setText(stringBuilder.toString());
     }
 
     @Override
@@ -233,7 +159,7 @@ public class IssuesToolWindowPanel extends SimpleToolWindowPanel implements ILis
         layout.show(mIssuesListPanel, "CardEMPTY");
 
         // Clear the summary
-        mIssueSummaryTextPane.setText("");
+        mIssueDetailsTextPane.setText("");
     }
 
     @Override
@@ -243,7 +169,7 @@ public class IssuesToolWindowPanel extends SimpleToolWindowPanel implements ILis
 
     @Override
     public void showDetailsPanel(boolean shouldShow) {
-        mSummaryPanel.setVisible(shouldShow);
+        mDetailsPanel.setVisible(shouldShow);
     }
 
     @Nullable
@@ -254,39 +180,28 @@ public class IssuesToolWindowPanel extends SimpleToolWindowPanel implements ILis
 
     @Override
     public boolean isDetailsPanelShown() {
-        return mSummaryPanel.isVisible();
+        return mDetailsPanel.isVisible();
     }
 
     /**
-     * Get a neat readable format of the date if it's not null, or an empty string
-     *
-     * @param date the date
-     * @return the date in a readable string format
+     * Initializes all actions
      */
-    @NotNull
-    private static String getValueOfDate(@Nullable Date date) {
-        String formattedDate = "";
-        if (date != null) {
-            formattedDate = DateFormatUtil.formatPrettyDateTime(date);
-        }
-        return formattedDate;
-    }
-
     private void initializeActions() {
         final AnAction refreshAction = ActionManager.getInstance().getAction(RefreshIssueListAction.ACTION_ID);
         final AnAction openIssueInBrowserAction = ActionManager.getInstance().getAction(OpenIssueInBrowserAction.ACTION_ID);
         final AnAction showDetailsPanelAction = ActionManager.getInstance().getAction(ShowDetailsPanelAction.ACTION_ID);
-
 
         final DefaultActionGroup actionGroup = new DefaultActionGroup();
         actionGroup.add(openIssueInBrowserAction);
         actionGroup.add(refreshAction);
         actionGroup.add(showDetailsPanelAction);
 
+        // Show actions in the toolbar
         final ActionToolbar actionToolbar = ActionManager.getInstance().createActionToolbar(ActionPlaces.UNKNOWN, actionGroup, false);
         actionToolbar.setTargetComponent(mToolbar);
         mToolbar.add(actionToolbar.getComponent());
 
+        // Show actions when selected issue is right clicked
         PopupHandler.installPopupHandler(mIssuesTable, actionGroup, ActionPlaces.UPDATE_POPUP, ActionManager.getInstance());
     }
 
@@ -305,10 +220,10 @@ public class IssuesToolWindowPanel extends SimpleToolWindowPanel implements ILis
         mIssuesTable.getSelectionModel().addListSelectionListener(e -> {
             final Task selectedIssue = mIssuesTable.getSelectedObject();
             if (selectedIssue != null) {
-                mPresenter.showSummary(selectedIssue);
+                mPresenter.showDetails(selectedIssue);
             }
         });
-        mIssueSummaryTextPane.addHyperlinkListener(new BrowserHyperlinkListener());
+        mIssueDetailsTextPane.addHyperlinkListener(new BrowserHyperlinkListener());
 
     }
